add OSC52 set clipboard support (#1421)
enable corresponding checkbox in to use it Options/Interface

+++ WinPort/src/APIConsole.cpp
@@ -318,7 +318,7 @@ extern "C" {
 		return TRUE;
 	}
 
-	WINPORT_DECL(SetConsoleTweaks, DWORD, (DWORD tweaks))
+	WINPORT_DECL(SetConsoleTweaks, DWORD64, (DWORD64 tweaks))
 	{
 		return g_winport_con_out->SetConsoleTweaks(tweaks);
 	}
